+++ packages/core/src/index.ts
@@ -1,4 +1,3 @@
-export * from './fp';
 export * from './guards';
 export * from './hooks';
 export * from './types';