+++ packages/forms/src/exceptions/validation-error-exception.test.ts
@@ -1,0 +1,11 @@
+import { ValidationException } from './validation-error.exception';
+
+describe('ValidationException', () => {
+  it('should properly detect validation error', () => {
+    expect(
+      ValidationException.isValidationError(new ValidationException([])),
+    ).toBe(true);
+
+    expect(ValidationException.isValidationError(new Error('abc'))).toBe(false);
+  });
+});