+++ packages/forms/src/index.ts
@@ -1,4 +1,2 @@
-export * from '@under-control/core';
-export * from '@under-control/inputs';
-export * from '@under-control/validate';
+export * from './exceptions';
 export * from './hooks';