+++ packages/validate/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @under-control/validate
 
+## 1.5.0
+
+### Minor Changes
+
+- Drop either support
+
+### Patch Changes
+
+- Updated dependencies
+  - @under-control/core@1.5.0
+
 ## 1.4.0
 
 ### Minor Changes